backend/apps/north_sea_watch/requirements/base.txt
# Django and REST framework
django>=4.2.0,<4.3.0
djangorestframework>=3.15.2,<3.16.0

# Database
dj-database-url>=2.1.0,<2.2.0
psycopg2-binary>=2.9.9,<2.10.0

# CORS
django-cors-headers>=4.3.1,<4.4.0

# Environment variables
django-environ>=0.11.2,<0.12.0

# Static files
whitenoise>=6.6.0,<6.7.0

# ABM Simulation
mesa==0.9.0
numpy
pandas
shapely
python-dateutil
pytz
scipy
six
tqdm
tzdata

# User Agent parser
user-agents>=2.2.0,<2.3.0

# HTTP requests
requests>=2.32.0